Module clippy_utils::ast_utils

Utilities for manipulating and extracting information from rustc_ast::ast.

The eq_foobar functions test for semantic equality but ignores NodeIds and Spans.

Re-exports

pub use ident_iter::IdentIter;

Modules

ident_iter

Functions

eq_angle_arg

eq_anon_const

eq_arm

eq_assoc_constraint

eq_assoc_item_kind

eq_attr

eq_block

eq_closure_binder

eq_defaultness

eq_expr

eq_expr_opt

eq_ext

eq_field

eq_field_pat

eq_fn_decl

eq_fn_header

eq_fn_ret_ty

eq_fn_sig

eq_foreign_item_kind

eq_generic_arg

eq_generic_args

eq_generic_bound

eq_generic_param

eq_generics

eq_id

eq_item

eq_item_kind

eq_label

eq_local_kind

eq_mac_args

eq_mac_call

eq_maybe_qself

eq_pat

eq_path

eq_path_seg

eq_poly_ref_trait

eq_qself

eq_range_end

eq_stmt

eq_str_lit

eq_struct_field

eq_struct_rest

eq_term 🔒

eq_ty

eq_use_tree

eq_use_tree_kind

eq_variant

eq_variant_data

eq_vis

eq_where_predicate

is_useless_with_eq_exprs

unordered_over

Checks if each element in the first slice is contained within the latter as per eq_fn.