Module clippy_utils::consts

Structs

ConstEvalLateContext

Enums

Constant

A LitKind-like enum to fold constant Exprs into.

ConstantSource

The source of a constant value.

FullInt

Traits

IntTypeBounds 🔒

Functions

constant

Attempts to evaluate the expression as a constant.

constant_full_int

constant_is_empty

Attempts to check whether the expression is a constant representing an empty slice, str, array, etc…

constant_simple

Attempts to evaluate an expression only if its value is not dependent on other items.

constant_with_source

Attempts to evaluate the expression as a constant.

field_of_struct 🔒

lit_to_mir_constant

Parses a LitKind to a Constant.

mir_is_empty 🔒

mir_to_const