Enum clippy_utils::consts::Constant

pub enum Constant {

    Str(String),
    Binary(Lrc<[u8]>),
    Char(char),
    Int(u128),
    F32(f32),
    F64(f64),
    Bool(bool),
    Vec(Vec<Constant>),
    Repeat(Box<Constant>, u64),
    Tuple(Vec<Constant>),
    RawPtr(u128),
    Ref(Box<Constant>),
    Err,

}

A LitKind-like enum to fold constant Exprs into.

Variants

Str(String)

A String (e.g., “abc”).

Binary(Lrc<[u8]>)

A binary string (e.g., b"abc").

Char(char)

A single char (e.g., 'a').

Int(u128)

An integer’s bit representation.

F32(f32)

An f32.

F64(f64)

An f64.

Bool(bool)

true or false.

Vec(Vec<Constant>)

An array of constants.

Repeat(Box<Constant>, u64)

Also an array, but with only one constant, repeated N times.

Tuple(Vec<Constant>)

A tuple of constants.

RawPtr(u128)

A raw pointer.

Ref(Box<Constant>)

A reference

Err

A literal with syntax error.

Implementations

impl Constant

pub fn partial_cmp(
tcx: TyCtxt<'_>,
cmp_type: Ty<'_>,
left: &Self,
right: &Self
) -> Option<Ordering>

pub fn int_value(&self, cx: &LateContext<'_>, val_type: Ty<'_>) -> Option<FullInt>

Returns the integer value or None if self or val_type is not integer type.

pub fn peel_refs(self) -> Self