Module clippy_utils::eager_or_lazy

Utilities for evaluating whether eagerly evaluated expressions can be made lazy and vice versa.

Things to consider:

does the expression have side-effects?

is the expression computationally expensive?

See lints:

unnecessary-lazy-evaluations

or-fun-call

option-if-let-else

Enums

EagernessSuggestion 🔒

Functions

expr_eagerness 🔒

fn_eagerness 🔒

Determine the eagerness of the given function call.

res_has_significant_drop 🔒

switch_to_eager_eval

Whether the given expression should be changed to evaluate eagerly

switch_to_lazy_eval

Whether the given expression should be changed to evaluate lazily