Enum clippy_utils::ExprUseNode

pub enum ExprUseNode<'tcx> {
    Local(&'tcx Local<'tcx>),
    ConstStatic(OwnerId),
    Return(OwnerId),
    Field(&'tcx ExprField<'tcx>),
    FnArg(&'tcx Expr<'tcx>, usize),
    MethodArg(HirId, Option<&'tcx GenericArgs<'tcx>>, usize),
    Callee,
    FieldAccess(Ident),
}

The node which consumes a value.

Variants

Local(&'tcx Local<'tcx>)

Assignment to, or initializer for, a local

ConstStatic(OwnerId)

Initializer for a const or static item.

Return(OwnerId)

Implicit or explicit return from a function.

Field(&'tcx ExprField<'tcx>)

Initialization of a struct field.

FnArg(&'tcx Expr<'tcx>, usize)

An argument to a function.

MethodArg(HirId, Option<&'tcx GenericArgs<'tcx>>, usize)

An argument to a method.

Callee

The callee of a function call.

FieldAccess(Ident)

Access of a field.

Implementations

impl<'tcx> ExprUseNode<'tcx>

pub fn is_return(&self) -> bool

Checks if the value is returned from the function.

pub fn is_recv(&self) -> bool

Checks if the value is used as a method call receiver.

pub fn defined_ty(&self, cx: &LateContext<'tcx>) -> Option<DefinedTy<'tcx>>

Gets the needed type as it’s defined without any type inference.