clippy_utils

Function is_mutable

pub fn is_mutable(cx: &LateContext<'_>, expr: &Expr<'_>) -> bool

Returns true if expr designates a mutable static, a mutable local binding, or an expression that can be owned.