Module clippy_utils::hir_utils

Structs

HirEqInterExpr

SpanlessEq

Type used to check whether two ast are the same. This is different from the operator == on ast types as this operator would compare true equality with ID and span.

SpanlessHash

Type used to hash an ast element. This is different from the Hash trait on ast types as this trait would consider IDs and spans.

Functions

both

Checks if the two Options are both None or some equal values as per eq_fn.

count_eq

Counts how many elements of the slices are equal as per eq_fn.

eq_expr_value

Checks if two expressions evaluate to the same value, and don’t contain any side effects.

hash_expr

hash_stmt

over

Checks if two slices are equal as per eq_fn.

reduce_exprkind 🔒

Some simple reductions like { return } => return

swap_binop 🔒

Type Definitions

SpanlessEqCallback 🔒

Callback that is called when two expressions are not equal in the sense of SpanlessEq, but other conditions would make them equal.