Struct clippy_utils::hir_utils::SpanlessEq

pub struct SpanlessEq<'a, 'tcx> {
    cx: &'a LateContext<'tcx>,
    maybe_typeck_results: Option<(&'tcx TypeckResults<'tcx>, &'tcx TypeckResults<'tcx>)>,
    allow_side_effects: bool,
    expr_fallback: Option<Box<dyn FnMut(&Expr<'_>, &Expr<'_>) -> bool + 'a>>,
}

Type used to check whether two ast are the same. This is different from the operator == on ast types as this operator would compare true equality with ID and span.

Note that some expressions kinds are not considered but could be added.

Fields
cx: &'a LateContext<'tcx>

Context used to evaluate constant expressions.

maybe_typeck_results: Option<(&'tcx TypeckResults<'tcx>, &'tcx TypeckResults<'tcx>)>allow_side_effects: boolexpr_fallback: Option<Box<dyn FnMut(&Expr<'_>, &Expr<'_>) -> bool + 'a>>
Implementations

impl<'a, 'tcx> SpanlessEq<'a, 'tcx>

pub fn new(cx: &'a LateContext<'tcx>) -> Self

pub fn deny_side_effects(self) -> Self

Consider expressions containing potential side effects as not equal.

pub fn expr_fallback(
self,
expr_fallback: impl FnMut(&Expr<'_>, &Expr<'_>) -> bool + 'a
) -> Self

pub fn inter_expr(&mut self) -> HirEqInterExpr<'_, 'a, 'tcx>

Use this method to wrap comparisons that may involve inter-expression context. See self.locals.

pub fn eq_block(&mut self, left: &Block<'_>, right: &Block<'_>) -> bool

pub fn eq_expr(&mut self, left: &Expr<'_>, right: &Expr<'_>) -> bool

pub fn eq_path(&mut self, left: &Path<'_>, right: &Path<'_>) -> bool

pub fn eq_path_segment(
&mut self,
left: &PathSegment<'_>,
right: &PathSegment<'_>
) -> bool

pub fn eq_path_segments(
&mut self,
left: &[PathSegment<'_>],
right: &[PathSegment<'_>]
) -> bool