clippy_utils::hir_utils
Type Alias SpanlessEqCallback

type SpanlessEqCallback<'a> = dyn FnMut(&Expr<'_>, &Expr<'_>) -> bool + 'a;

Callback that is called when two expressions are not equal in the sense of SpanlessEq, but other conditions would make them equal.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: (unsized)