clippy_utils

Module msrvs

Macros

msrv_aliases 🔒

Structs

Msrv

Tracks the current MSRV from clippy.toml, Cargo.toml or set via #[clippy::msrv] in late lint passes, use MsrvStack for early passes

MsrvStack

Tracks the current MSRV from clippy.toml, Cargo.toml or set via #[clippy::msrv] in early lint passes, use Msrv for late passes

Constants

ABS_DIFF

ARRAY_IMPL_ANY_LEN

ARRAY_INTO_ITERATOR

BINARY_HEAP_RETAIN

BOOL_THEN

BOOL_THEN_SOME

BORROW_AS_PTR

BOX_INTO_ITER

BTREE_MAP_RETAIN

BTREE_SET_RETAIN

BUILD_HASHER_HASH_ONE

CLAMP

CLONE_INTO

CONST_CHAR_IS_DIGIT

CONST_EXTERN_C_FN

CONST_EXTERN_FN

CONST_FLOAT_BITS_CONV

CONST_FLOAT_CLASSIFY

CONST_FN_TRAIT_BOUND

CONST_FN_UNION

CONST_IF_MATCH

CONST_IS_POWER_OF_TWO

CONST_MUT_REFS

CONST_OPTION_AS_SLICE

CONST_RAW_PTR_DEREF

CONST_SIZE_OF_VAL

CONST_SLICE_FROM_REF

CONST_UNWRAP

C_STR_LITERALS

DEFAULT_ENUM_ATTRIBUTE

DIV_CEIL

DURATION_ABS_DIFF

ERROR_IN_CORE

EXPECT_ERR

EXPLICIT_SELF_TYPE_ELISION

FIELD_INIT_SHORTHAND

FORMAT_ARGS_CAPTURE

FROM_BOOL

HASH_MAP_RETAIN

HASH_SET_RETAIN

INTEGER_BITS

INTO_KEYS

INT_MIDPOINT

IO_ERROR_OTHER

IS_ASCII_DIGIT

IS_ASCII_DIGIT_CONST

IS_NONE_OR

ITERATOR_COPIED

ITERATOR_FIND_MAP

ITERATOR_TRY_FOLD

ITER_FLATTEN

LAZY_CELL

LET_CHAINS

LET_ELSE

LINT_REASONS_STABILIZATION

LOG2_10

LOG10_2

MANUAL_DANGLING_PTR

MAP_WHILE

MATCHES_MACRO

MAYBE_BOUND_IN_WHERE

MEM_TAKE

NON_EXHAUSTIVE

NUMERIC_ASSOCIATED_CONSTANTS

OPTION_AS_DEREF

OPTION_AS_SLICE

OPTION_COPIED

OPTION_REPLACE

OPTION_RESULT_INSPECT

OPTION_RESULT_IS_VARIANT_AND

OR_PATTERNS

OS_STR_DISPLAY

PATH_MAIN_SEPARATOR_STR

PATTERN_TRAIT_CHAR_ARRAY

POINTER_CAST

POINTER_CAST_CONSTNESS

PTR_FROM_REF

PTR_NULL

PTR_SLICE_RAW_PARTS

QUESTION_MARK_OPERATOR

RANGE_CONTAINS

RANGE_INCLUSIVE

RAW_REF_OP

REM_EUCLID

REM_EUCLID_CONST

REPEAT_N

REPEAT_WITH

REPR_RUST

RESULT_MAP_OR_ELSE

RE_REBALANCING_COHERENCE

SATURATING_SUB_CONST

SEEK_FROM_CURRENT

SEEK_REWIND

SLICE_FILL

SLICE_FROM_REF

SLICE_PATTERNS

STATIC_IN_CONST

STRING_RETAIN

STR_REPEAT

STR_SPLIT_ONCE

STR_STRIP_PREFIX

TAU

THREAD_LOCAL_CONST_INIT

TOOL_ATTRIBUTES

TRY_FROM

TUPLE_ARRAY_CONVERSIONS

TYPE_ALIAS_ENUM_VARIANTS

UINT_FLOAT_MIDPOINT

UNDERSCORE_IMPORTS

UNSIGNED_ABS

UNSIGNED_IS_MULTIPLE_OF

Statics

SEEN_MSRV_ATTR 🔒

#[clippy::msrv] attributes are rarely used outside of Clippy’s test suite, as a basic optimization we can skip traversing the HIR in Msrv::meets if we never saw an MSRV attribute during the early lint passes

Functions

parse_attrs 🔒