Module clippy_utils::numeric_literal

Structs

NumericLiteral

Enums

Radix

Functions

format

A helper method to format numeric literals with digit grouping. lit must be a valid numeric literal without suffix.

lit_suffix_length 🔒

split_suffix 🔒