Module clippy_utils::paths

This module contains paths to types and functions Clippy needs to know about.

Whenever possible, please consider diagnostic items over hardcoded paths. See https://github.com/rust-lang/rust-clippy/issues/5393 for more information.

Constants

ARC_PTR_EQ

BTREEMAP_CONTAINS_KEY

BTREEMAP_INSERT

BTREESET_ITER

CLONE_TRAIT_METHOD

CONVERT_IDENTITY

CORE_ITER_CLONED

CORE_ITER_COLLECT

CORE_ITER_COPIED

CORE_ITER_FILTER

CSTRING_AS_C_STR

DEFAULT_TRAIT_METHOD

DEREF_MUT_TRAIT_METHOD

EXIT

F32_EPSILON

F64_EPSILON

FROM_ITERATOR_METHOD

FROM_STR_METHOD

FUTURES_IO_ASYNCREADEXT

FUTURES_IO_ASYNCWRITEEXT

HASHMAP_CONTAINS_KEY

HASHMAP_INSERT

HASHSET_ITER

INSERT_STR

INSTANT

INSTANT_NOW

ITERTOOLS_NEXT_TUPLE

ITER_COUNT

ITER_EMPTY

MEM_SWAP

OPEN_OPTIONS

OS_STRING_AS_OS_STR

OS_STR_TO_OS_STRING

PARKING_LOT_MUTEX_GUARD

PARKING_LOT_RWLOCK_READ_GUARD

PARKING_LOT_RWLOCK_WRITE_GUARD

PATH_BUF_AS_PATH

PATH_TO_PATH_BUF

PEEKABLE

PERMISSIONS

PERMISSIONS_FROM_MODE

PTR_COPY

PTR_COPY_NONOVERLAPPING

PTR_EQ

PTR_NON_NULL

PTR_READ

PTR_READ_UNALIGNED

PTR_READ_VOLATILE

PTR_REPLACE

PTR_SLICE_FROM_RAW_PARTS

PTR_SLICE_FROM_RAW_PARTS_MUT

PTR_SWAP

PTR_SWAP_NONOVERLAPPING

PTR_UNALIGNED_VOLATILE_LOAD

PTR_UNALIGNED_VOLATILE_STORE

PTR_WRITE

PTR_WRITE_BYTES

PTR_WRITE_UNALIGNED

PTR_WRITE_VOLATILE

PUSH_STR

RANGE_ARGUMENT_TRAIT

RC_PTR_EQ

REFCELL_REF

REFCELL_REFMUT

REGEX_BUILDER_NEW

REGEX_BYTES_BUILDER_NEW

REGEX_BYTES_NEW

REGEX_BYTES_SET_NEW

REGEX_NEW

REGEX_SET_NEW

SERDE_DESERIALIZE

SERDE_DE_VISITOR

SLICE_FROM_RAW_PARTS

SLICE_FROM_RAW_PARTS_MUT

SLICE_GET

SLICE_INTO

SLICE_INTO_VEC

SLICE_ITER

STDERR

STDOUT

STD_FS_CREATE_DIR

STD_IO_SEEK

STD_IO_SEEKFROM_START

STD_IO_SEEK_FROM_CURRENT

STRING_AS_MUT_STR

STRING_AS_STR

STRING_NEW

STR_BYTES

STR_CHARS

STR_ENDS_WITH

STR_FROM_UTF8

STR_FROM_UTF8_UNCHECKED

STR_LEN

STR_STARTS_WITH

TOKIO_IO_ASYNCREADEXT

TOKIO_IO_ASYNCWRITEEXT

TO_OWNED_METHOD

TO_STRING_METHOD

TRY_FROM

VEC_AS_MUT_SLICE

VEC_AS_SLICE

VEC_DEQUE_ITER

VEC_FROM_ELEM

VEC_NEW

VEC_RESIZE

WEAK_ARC

WEAK_RC