clippy_utils::paths

Function lookup_path

pub fn lookup_path(tcx: TyCtxt<'_>, ns: PathNS, path: &[Symbol]) -> Vec<DefId>

Resolves a def path like std::vec::Vec.

Typically it will return one DefId or none, but in some situations there can be multiple:

memchr::memchr could return the functions from both memchr 1.0 and memchr 2.0

alloc::boxed::Box::downcast would return a function for each of the different inherent impls (1, 2, 3)

This function is expensive and should be used sparingly.