clippy_utils::paths

Struct PathLookup

pub struct PathLookup {
    ns: PathNS,
    path: &'static [Symbol],
    once: OnceLock<Vec<DefId>>,
}

Lazily resolves a path into a list of DefIds using lookup_path.

Typically it will contain one DefId or none, but in some situations there can be multiple:

memchr::memchr could return the functions from both memchr 1.0 and memchr 2.0

alloc::boxed::Box::downcast would return a function for each of the different inherent impls (1, 2, 3)

Fields
ns: PathNSpath: &'static [Symbol]once: OnceLock<Vec<DefId>>
Implementations

impl PathLookup

pub fn get(&self, cx: &LateContext<'_>) -> &[DefId]

Returns the list of DefIds that the path resolves to

pub fn only(&self, cx: &LateContext<'_>) -> Option<DefId>

Returns the single DefId that the path resolves to, this can only be used for paths into stdlib crates to avoid the issue of multiple DefIds being returned

May return None in no_std/no_core environments

pub fn matches(&self, cx: &LateContext<'_>, def_id: DefId) -> bool

Checks if the path resolves to the given def_id

pub fn matches_path<'tcx>( &self, cx: &LateContext<'_>, maybe_path: &impl MaybePath<'tcx>, ) -> bool

Resolves maybe_path to a DefId and checks if the PathLookup matches it

pub fn matches_ty(&self, cx: &LateContext<'_>, ty: Ty<'_>) -> bool

Checks if the path resolves to ty’s definition, must be an Adt