Function clippy_utils::sugg::deref_closure_args

pub fn deref_closure_args<'tcx>(
    cx: &LateContext<'_>,
    closure: &'tcx Expr<'_>
) -> Option<DerefClosure>

Build suggestion gradually by handling closure arg specific usages, such as explicit deref and borrowing cases. Returns None if no such use cases have been triggered in closure body

note: this only works on single line immutable closures with exactly one input parameter.