clippy_utils::ty::type_certainty
Function qpath_certainty

fn qpath_certainty(
    cx: &LateContext<'_>,
    qpath: &QPath<'_>,
    resolves_to_type: bool,
) -> Certainty

Tries to tell whether a QPath resolves to something certain, e.g., whether all of its path segments generic arguments are instantiated.

qpath could refer to either a type or a value. The heuristic never needs the DefId of a value. So DefIds are retained only when resolves_to_type is true.