clippy_utils

Module usage

Structs

BindingUsageFinder

MutVarsDelegate 🔒

ParamBindingIdCollector

Functions

contains_return_break_continue_macro

contains_todo_unimplement_macro

Checks if the given expression contains macro call to todo!() or unimplemented!().

is_potentially_local_place

is_potentially_mutated

is_todo_unimplemented_macro

Checks if the given expression is a macro call to todo!() or unimplemented!().

is_todo_unimplemented_stub

Checks if the given expression is a stub, i.e., a todo!() or unimplemented!() expression, or a block whose last expression is a todo!() or unimplemented!().

local_used_after_expr

local_used_in

mutated_variables

Returns a set of mutated local variable IDs, or None if mutations could not be determined.