clippy_utils::consts
Enum Constant

pub enum Constant<'tcx> {

    Adt(Const<'tcx>),
    Str(String),
    Binary(Lrc<[u8]>),
    Char(char),
    Int(u128),
    F16(f16),
    F32(f32),
    F64(f64),
    F128(f128),
    Bool(bool),
    Vec(Vec<Constant<'tcx>>),
    Repeat(Box<Constant<'tcx>>, u64),
    Tuple(Vec<Constant<'tcx>>),
    RawPtr(u128),
    Ref(Box<Constant<'tcx>>),
    Err,

}

A LitKind-like enum to fold constant Exprs into.

Variants

Adt(Const<'tcx>)

Str(String)

A String (e.g., โabcโ).

Binary(Lrc<[u8]>)

A binary string (e.g., b"abc").

Char(char)

A single char (e.g., 'a').

Int(u128)

An integerโs bit representation.

F16(f16)

An f16.

F32(f32)

An f32.

F64(f64)

An f64.

F128(f128)

An f128.

Bool(bool)

true or false.

Vec(Vec<Constant<'tcx>>)

An array of constants.

Repeat(Box<Constant<'tcx>>, u64)

Also an array, but with only one constant, repeated N times.

Tuple(Vec<Constant<'tcx>>)

A tuple of constants.

RawPtr(u128)

A raw pointer.

Ref(Box<Constant<'tcx>>)

A reference

Err

A literal with syntax error.

Implementations

impl Constant<'_>

pub fn partial_cmp( tcx: TyCtxt<'_>, cmp_type: Ty<'_>, left: &Self, right: &Self, ) -> Option<Ordering>

pub fn int_value(&self, tcx: TyCtxt<'_>, val_type: Ty<'_>) -> Option<FullInt>

Returns the integer value or None if self or val_type is not integer type.

pub fn peel_refs(self) -> Self

fn parse_f16(s: &str) -> Self

fn parse_f128(s: &str) -> Self