clippy_utils
Module diagnostics

Clippy wrappers around rustc’s diagnostic functions.

These functions are used by the INTERNAL_METADATA_COLLECTOR lint to collect the corresponding lint applicability. Please make sure that you update the LINT_EMISSION_FUNCTIONS variable in clippy_lints::utils::internal_lints::metadata_collector when a new function is added or renamed.

Thank you! ~The INTERNAL_METADATA_COLLECTOR lint

Functions

docs_link 🔒

span_lint

Emit a basic lint message with a msg and a span.

span_lint_and_help

Same as span_lint but with an extra help message.

span_lint_and_note

Like span_lint but with a note section instead of a help message.

span_lint_and_sugg

Add a span lint with a suggestion on how to fix it.

span_lint_and_then

Like span_lint but allows to add notes, help and suggestions using a closure.

span_lint_hir

Like span_lint, but emits the lint at the node identified by the given HirId.

span_lint_hir_and_then

Like span_lint_and_then, but emits the lint at the node identified by the given HirId.

validate_diag 🔒

Makes sure that a diagnostic is well formed.