clippy_utils
Module higher

This module contains functions that retrieve specific elements.

Structs

ForLoop

The essential nodes of a desugared for loop as well as the entire span: for pat in arg { body } becomes (pat, arg, body). Returns (pat, arg, body, span).

If

An if expression without DropTemps

IfLet

An if let expression

IfOrIfLet

An if or if let expression

Range

Represent a range akin to ast::ExprKind::Range.

While

A desugared while loop

WhileLet

A desugared while let loop

Enums

IfLetOrMatch

An if let or match expression. Useful for lints that trigger on one or the other.

VecArgs

Represents the pre-expansion arguments of a vec! invocation.

VecInitKind

A parsed Vec initialization expression

Functions

binop

Converts a hir binary operator to the corresponding ast type.

get_vec_init_kind

Checks if the given expression is an initialization of Vec and returns its kind.