Struct clippy_utils::hir_utils::SpanlessHash

pub struct SpanlessHash<'a, 'tcx> {
    cx: &'a LateContext<'tcx>,
    maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>,
    s: FxHasher,
}

Type used to hash an ast element. This is different from the Hash trait on ast types as this trait would consider IDs and spans.

All expressions kind are hashed, but some might have a weaker hash.

Fields
cx: &'a LateContext<'tcx>

Context used to evaluate constant expressions.

maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>s: FxHasher
Implementations

impl<'a, 'tcx> SpanlessHash<'a, 'tcx>

pub fn new(cx: &'a LateContext<'tcx>) -> Self

pub fn finish(self) -> u64

pub fn hash_block(&mut self, b: &Block<'_>)

pub fn hash_expr(&mut self, e: &Expr<'_>)

pub fn hash_exprs(&mut self, e: &[Expr<'_>])

pub fn hash_name(&mut self, n: Symbol)

pub fn hash_qpath(&mut self, p: &QPath<'_>)

pub fn hash_pat(&mut self, pat: &Pat<'_>)

pub fn hash_path(&mut self, path: &Path<'_>)

pub fn hash_stmt(&mut self, b: &Stmt<'_>)

pub fn hash_guard(&mut self, g: &Guard<'_>)

pub fn hash_lifetime(&mut self, lifetime: &Lifetime)

pub fn hash_ty(&mut self, ty: &Ty<'_>)

pub fn hash_tykind(&mut self, ty: &TyKind<'_>)

pub fn hash_array_length(&mut self, length: ArrayLen)

pub fn hash_body(&mut self, body_id: BodyId)

fn hash_generic_args(&mut self, arg_list: &[GenericArg<'_>])