clippy_utils
Module mir

Modules

possible_borrower 🔒

possible_origin 🔒

transitive_relation 🔒

Structs

LocalUsage

PossibleBorrowerMap

Result of PossibleBorrowerVisitor.

V 🔒

Functions

block_in_cycle

Checks if the block is part of a cycle

enclosing_mir

Returns the mir::Body containing the node associated with hir_id.

expr_local

Tries to determine the Local corresponding to expr, if any. This function is expensive and should be used sparingly.

is_local_assignment 🔒

local_assignments

Returns a vector of mir::Location where local is assigned.

used_exactly_once

Convenience wrapper around visit_local_usage.

visit_local_usage