clippy_utils
Module paths

This module contains paths to types and functions Clippy needs to know about.

Whenever possible, please consider diagnostic items over hardcoded paths. See https://github.com/rust-lang/rust-clippy/issues/5393 for more information.

Constants

ABORT

APPLICABILITY

APPLICABILITY_VALUES

CHILD

CHILD_ID

CHILD_KILL

DIAG

EARLY_CONTEXT

EARLY_LINT_PASS

FUTURES_IO_ASYNCREADEXT

FUTURES_IO_ASYNCWRITEEXT

IDENT

IDENT_AS_STR

ITERTOOLS_NEXT_TUPLE

KW_MODULE

LATE_CONTEXT

LATE_LINT_PASS

LINT

MSRV

PANIC_ANY

PARKING_LOT_MUTEX_GUARD

PARKING_LOT_RWLOCK_READ_GUARD

PARKING_LOT_RWLOCK_WRITE_GUARD

REGEX_BUILDER_NEW

REGEX_BYTES_BUILDER_NEW

REGEX_BYTES_NEW

REGEX_BYTES_SET_NEW

REGEX_NEW

REGEX_SET_NEW

SERDE_DESERIALIZE

SERDE_DE_VISITOR

SYMBOL

SYMBOL_AS_STR

SYMBOL_INTERN

SYMBOL_TO_IDENT_STRING

SYM_MODULE

SYNTAX_CONTEXT

TOKIO_FILE_OPTIONS

TOKIO_IO_ASYNCREADEXT

TOKIO_IO_ASYNCWRITEEXT

TOKIO_IO_OPEN_OPTIONS

TOKIO_IO_OPEN_OPTIONS_NEW