clippy_utils

Module paths

This module contains paths to types and functions Clippy needs to know about.

Whenever possible, please consider diagnostic items over hardcoded paths. See https://github.com/rust-lang/rust-clippy/issues/5393 for more information.

Macros

path_macros 🔒

Structs

PathLookup

Lazily resolves a path into a list of DefIds using lookup_path.

Enums

PathNS

Specifies whether to resolve a path in the TypeNS, ValueNS, MacroNS or in an arbitrary namespace

Statics

FUTURES_IO_ASYNCREADEXT

FUTURES_IO_ASYNCWRITEEXT

ITERTOOLS_NEXT_TUPLE

LAZY_STATIC

ONCE_CELL_SYNC_LAZY

ONCE_CELL_SYNC_LAZY_NEW

PARKING_LOT_GUARDS

REGEX_BUILDER_NEW

REGEX_BYTES_BUILDER_NEW

REGEX_BYTES_NEW

REGEX_BYTES_SET_NEW

REGEX_NEW

REGEX_SET_NEW

SERDE_DESERIALIZE

SERDE_DE_VISITOR

TOKIO_FILE_OPTIONS

TOKIO_IO_ASYNCREADEXT

TOKIO_IO_ASYNCWRITEEXT

TOKIO_IO_OPEN_OPTIONS

TOKIO_IO_OPEN_OPTIONS_NEW

Functions

find_crates

Finds the crates called name, may be multiple due to multiple major versions.

find_primitive_impls 🔒

item_child_by_name 🔒

local_item_child_by_name 🔒

lookup_path

Resolves a def path like std::vec::Vec.

lookup_path_str

Equivalent to a lookup_path after splitting the input string on ::

lookup_with_base 🔒

Resolves a def path like vec::Vec with the base std.

non_local_item_child_by_name 🔒