Module clippy_utils::paths

This module contains paths to types and functions Clippy needs to know about.

Whenever possible, please consider diagnostic items over hardcoded paths. See https://github.com/rust-lang/rust-clippy/issues/5393 for more information.

Constants

APPLICABILITY

APPLICABILITY_VALUES

BINARYHEAP_ITER

BOOL_THEN

BTREEMAP_CONTAINS_KEY

BTREEMAP_INSERT

BTREESET_ITER

CLONE_TRAIT_METHOD

CORE_ITER_CLONED

CORE_ITER_COPIED

CORE_ITER_FILTER

CORE_RESULT_OK_METHOD

CSTRING_AS_C_STR

DIAG

EARLY_CONTEXT

EARLY_LINT_PASS

F32_EPSILON

F64_EPSILON

FILE_OPTIONS

FUTURES_IO_ASYNCREADEXT

FUTURES_IO_ASYNCWRITEEXT

HASHMAP_CONTAINS_KEY

HASHMAP_DRAIN

HASHMAP_INSERT

HASHMAP_ITER

HASHMAP_ITER_MUT

HASHMAP_KEYS

HASHMAP_VALUES

HASHMAP_VALUES_MUT

HASHSET_DRAIN

HASHSET_ITER

HASHSET_ITER_TY

IDENT

IDENT_AS_STR

INSERT_STR

INSTANT_NOW

ITERTOOLS_NEXT_TUPLE

KW_MODULE

LATE_CONTEXT

LATE_LINT_PASS

LINT

MSRV

OPEN_OPTIONS_NEW

OPTION_EXPECT

OPTION_UNWRAP

OS_STRING_AS_OS_STR

OS_STR_TO_OS_STRING

PARKING_LOT_MUTEX_GUARD

PARKING_LOT_RWLOCK_READ_GUARD

PARKING_LOT_RWLOCK_WRITE_GUARD

PATH_BUF_AS_PATH

PATH_MAIN_SEPARATOR

PATH_TO_PATH_BUF

PERMISSIONS_FROM_MODE

PUSH_STR

REGEX_BUILDER_NEW

REGEX_BYTES_BUILDER_NEW

REGEX_BYTES_NEW

REGEX_BYTES_SET_NEW

REGEX_NEW

REGEX_SET_NEW

SERDE_DESERIALIZE

SERDE_DE_VISITOR

SLICE_INTO

SLICE_INTO_VEC

STD_IO_SEEKFROM_START

STD_IO_SEEK_FROM_CURRENT

STRING_AS_MUT_STR

STRING_AS_STR

STRING_NEW

STR_CHARS

STR_ENDS_WITH

STR_LEN

STR_STARTS_WITH

SYMBOL

SYMBOL_AS_STR

SYMBOL_INTERN

SYMBOL_TO_IDENT_STRING

SYM_MODULE

SYNTAX_CONTEXT

TOKIO_FILE_OPTIONS

TOKIO_IO_ASYNCREADEXT

TOKIO_IO_ASYNCWRITEEXT

TOKIO_IO_OPEN_OPTIONS

TOKIO_IO_OPEN_OPTIONS_NEW

VEC_AS_MUT_SLICE

VEC_AS_SLICE

VEC_DEQUE_ITER

VEC_FROM_ELEM

VEC_IS_EMPTY

VEC_NEW

VEC_POP

VEC_WITH_CAPACITY

WAKER