Module clippy_utils::paths

This module contains paths to types and functions Clippy needs to know about.

Whenever possible, please consider diagnostic items over hardcoded paths. See https://github.com/rust-lang/rust-clippy/issues/5393 for more information.

Constants

BINARYHEAP_ITER

BOOL_THEN

BTREEMAP_CONTAINS_KEY

BTREEMAP_INSERT

BTREESET_ITER

CLONE_TRAIT_METHOD

CORE_ITER_CLONED

CORE_ITER_COPIED

CORE_ITER_FILTER

CORE_RESULT_OK_METHOD

CSTRING_AS_C_STR

F32_EPSILON

F64_EPSILON

FUTURES_IO_ASYNCREADEXT

FUTURES_IO_ASYNCWRITEEXT

HASHMAP_CONTAINS_KEY

HASHMAP_INSERT

HASHSET_ITER

INSERT_STR

INSTANT_NOW

ITERTOOLS_NEXT_TUPLE

OPTION_EXPECT

OPTION_UNWRAP

OS_STRING_AS_OS_STR

OS_STR_TO_OS_STRING

PARKING_LOT_MUTEX_GUARD

PARKING_LOT_RWLOCK_READ_GUARD

PARKING_LOT_RWLOCK_WRITE_GUARD

PATH_BUF_AS_PATH

PATH_MAIN_SEPARATOR

PATH_TO_PATH_BUF

PERMISSIONS_FROM_MODE

PUSH_STR

REGEX_BUILDER_NEW

REGEX_BYTES_BUILDER_NEW

REGEX_BYTES_NEW

REGEX_BYTES_SET_NEW

REGEX_NEW

REGEX_SET_NEW

SERDE_DESERIALIZE

SERDE_DE_VISITOR

SLICE_GET

SLICE_INTO

SLICE_INTO_VEC

STD_IO_SEEKFROM_START

STD_IO_SEEK_FROM_CURRENT

STRING_AS_MUT_STR

STRING_AS_STR

STRING_NEW

STR_BYTES

STR_CHARS

STR_ENDS_WITH

STR_LEN

STR_STARTS_WITH

TOKIO_IO_ASYNCREADEXT

TOKIO_IO_ASYNCWRITEEXT

TO_STRING_METHOD

VEC_AS_MUT_SLICE

VEC_AS_SLICE

VEC_DEQUE_ITER

VEC_FROM_ELEM

VEC_IS_EMPTY

VEC_NEW

VEC_POP

VEC_RESIZE

VEC_WITH_CAPACITY