clippy_utils

Module str_utils

Structs

StrCount

Dealing with string comparison can be complicated, this struct ensures that both the character and byte count are provided for correct indexing.

StrIndex

Dealing with string indices can be hard, this struct ensures that both the character and byte index are provided for correct indexing.

Functions

camel_case_indices

Get the indexes of camel case components of a string s

camel_case_split

Split camel case string into a vector of its components

camel_case_start

Returns index of the first camel-case component of s.

camel_case_start_from_idx

Returns StrIndex of the last camel-case component of s[idx..].

camel_case_until

Returns the index of the character after the first camel-case component of s.

count_match_end

Returns the number of chars and bytes that match from the end

count_match_start

Returns the number of chars that match from the start

to_camel_case

Returns a CamelCase version of the input

to_snake_case

Returns a snake_case version of the input