clippy_utils::sugg

Enum Sugg

pub enum Sugg<'a> {
    NonParen(Cow<'a, str>),
    MaybeParen(Cow<'a, str>),
    BinOp(AssocOp, Cow<'a, str>, Cow<'a, str>),
    UnOp(UnOp, Box<Sugg<'a>>),
}

A helper type to build suggestion correctly handling parentheses.

Variants

NonParen(Cow<'a, str>)

An expression that never needs parentheses such as 1337 or [0; 42].

MaybeParen(Cow<'a, str>)

An expression that does not fit in other variants.

BinOp(AssocOp, Cow<'a, str>, Cow<'a, str>)

A binary operator expression, including as-casts and explicit type coercion.

UnOp(UnOp, Box<Sugg<'a>>)

A unary operator expression. This is used to sometimes represent ! or -, but only if the type with and without the operator is kept identical. It means that doubling the operator can be used to remove it instead, in order to provide better suggestions.

Implementations

impl<'a> Sugg<'a>

pub fn hir_opt(cx: &LateContext<'_>, expr: &Expr<'_>) -> Option<Self>

Prepare a suggestion from an expression.

pub fn hir(cx: &LateContext<'_>, expr: &Expr<'_>, default: &'a str) -> Self

Convenience function around hir_opt for suggestions with a default text.

pub fn hir_with_applicability( cx: &LateContext<'_>, expr: &Expr<'_>, default: &'a str, applicability: &mut Applicability, ) -> Self

Same as hir, but it adapts the applicability level by following rules:

Applicability level Unspecified will never be changed.

If the span is inside a macro, change the applicability level to MaybeIncorrect.

If the default value is used and the applicability level is MachineApplicable, change it to HasPlaceholders

pub fn hir_with_context( cx: &LateContext<'_>, expr: &Expr<'_>, ctxt: SyntaxContext, default: &'a str, applicability: &mut Applicability, ) -> Self

Same as hir, but first walks the span up to the given context. This will result in the macro call, rather than the expansion, if the span is from a child context. If the span is not from a child context, it will be used directly instead.

e.g. Given the expression &vec![], getting a snippet from the span for vec![] as a HIR node would result in box []. If given the context of the address of expression, this function will correctly get a snippet of vec![].

fn hir_from_snippet( expr: &Expr<'_>, get_snippet: impl FnMut(Span) -> Cow<'a, str>, ) -> Self

Generate a suggestion for an expression with the given snippet. This is used by the hir_* function variants of Sugg, since these use different snippet functions.

pub fn ast( cx: &EarlyContext<'_>, expr: &Expr, default: &'a str, ctxt: SyntaxContext, app: &mut Applicability, ) -> Self

Prepare a suggestion from an expression.

pub fn and(self, rhs: &Self) -> Sugg<'static>

Convenience method to create the <lhs> && <rhs> suggestion.

pub fn bit_and(self, rhs: &Self) -> Sugg<'static>

Convenience method to create the <lhs> & <rhs> suggestion.

pub fn as_ty<R: Display>(self, rhs: R) -> Sugg<'static>

Convenience method to create the <lhs> as <rhs> suggestion.

pub fn addr(self) -> Sugg<'static>

Convenience method to create the &<expr> suggestion.

pub fn mut_addr(self) -> Sugg<'static>

Convenience method to create the &mut <expr> suggestion.

pub fn deref(self) -> Sugg<'static>

Convenience method to create the *<expr> suggestion.

pub fn addr_deref(self) -> Sugg<'static>

Convenience method to create the &*<expr> suggestion. Currently this is needed because sugg.deref().addr() produces an unnecessary set of parentheses around the deref.

pub fn mut_addr_deref(self) -> Sugg<'static>

Convenience method to create the &mut *<expr> suggestion. Currently this is needed because sugg.deref().mut_addr() produces an unnecessary set of parentheses around the deref.

pub fn make_return(self) -> Sugg<'static>

Convenience method to transform suggestion into a return call

pub fn blockify(self) -> Sugg<'static>

Convenience method to transform suggestion into a block where the suggestion is a trailing expression

pub fn asyncify(self) -> Sugg<'static>

Convenience method to prefix the expression with the async keyword. Can be used after blockify to create an async block.

pub fn range(self, end: &Self, limits: RangeLimits) -> Sugg<'static>

Convenience method to create the <lhs>..<rhs> or <lhs>...<rhs> suggestion.

pub fn maybe_paren(self) -> Self

Adds parentheses to any expression that might need them. Suitable to the self argument of a method call (e.g., to build bar.foo() or (1 + 2).foo()).

pub fn into_string(self) -> String

fn starts_with_unary_op(&self) -> bool

Checks if self starts with a unary operator.

fn maybe_inner_paren(self) -> Self

Call maybe_paren on self if it doesn’t start with a unary operator, don’t touch it otherwise.